requirements.txt
pennylane==0.35.1
torch==2.2.0
numpy==1.26.4
matplotlib==3.8.4
jupyterlab==4.1.6
torch_geometric==2.5.3
comet_ml==3.41.0
python-dotenv==1.0.1
torch_scatter==2.1.2
torch_sparse==0.6.18